bitflag_attr::example_generated
Struct ExampleFlags

pub struct ExampleFlags(/* private fields */)
where
    u32: BitsPrimitive;

A example bitflag

Implementations

impl ExampleFlags

pub const Flag1: Self

pub const Flag2: Self

pub const Flag3: Self

pub const Flag4: Self

pub const Flag5: Self

pub const Flag6: Self

pub const Flag7: Self

pub const Flag8: Self

pub const Flag9: Self

impl ExampleFlags

pub const fn bits(&self) -> u32

Return the underlying bits of this bitflag.

pub const fn from_bits(bits: u32) -> Option<Self>

Converts from a bits value. Returning None is any unknown bits are set.

pub const fn from_bits_truncate(bits: u32) -> Self

Convert from bits value, unsetting any unknown bits.

pub const fn from_bits_retain(bits: u32) -> Self

Convert from bits value exactly.

pub fn from_flag_name(name: &str) -> Option<Self>

Convert from a flag name.

pub const fn empty() -> Self

Construct a flags value with all bits unset.

pub const fn is_empty(&self) -> bool

Returns true if the flag value has all bits unset.

pub const fn all_bits() -> Self

Returns a flag value that contains all value.

This will include bits that do not have any flags/meaning. Use all if you want only the specified flags set.

pub const fn is_all_bits(&self) -> bool

Returns true if the flag value contains all value bits set.

This will check for all bits. Use is_all if you want to check for all specified flags.

pub const fn all() -> Self

Construct a flag value with all known flags set.

This will only set the flags specified as associated constant.

pub const fn is_all(&self) -> bool

Returns true if the flag value contais all known flags.

pub const fn contains_unknown_bits(&self) -> bool

Returns true if there are any unknown bits set in the flag value.

pub const fn truncated(&self) -> Self

Returns a bit flag that only has bits corresponding to the specified flags as associated constant.

pub fn truncate(&mut self)

Removes unknown bits from the flag value.

pub const fn intersects(&self, other: Self) -> bool

Returns true if this flag value intersects with any value in other.

This is equivalent to (self & other) != Self::empty()

pub const fn contains(&self, other: Self) -> bool

Returns true if this flag value contains all values of other.

This is equivalent to (self & other) == other

pub const fn not(self) -> Self

Returns the bitwise NOT of the flag.

This function does not truncate unused bits (bits that do not have any flags/meaning). Use complement if you want that the result to be truncated in one call.

pub const fn and(self, other: Self) -> Self

Returns the bitwise AND of the flag.

pub const fn or(self, other: Self) -> Self

Returns the bitwise OR of the flag with other.

pub const fn xor(self, other: Self) -> Self

Returns the bitwise XOR of the flag with other.

pub const fn intersection(self, other: Self) -> Self

Returns the intersection from this value with other.

pub const fn union(self, other: Self) -> Self

Returns the union from this value with other.

pub const fn difference(self, other: Self) -> Self

Returns the difference from this value with other.

In other words, returns the intersection of this value with the negation of other.

This method is not equivalent to self & !other when other has unknown bits set. difference won’t truncate other, but the ! operator will.

pub const fn symmetric_difference(self, other: Self) -> Self

Returns the symmetric difference from this value with other.

pub const fn complement(self) -> Self

Returns the complement of the value.

This is very similar to the not, but truncates non used bits.

pub fn set(&mut self, other: Self)

Set the flags in other in the value.

pub fn unset(&mut self, other: Self)

Unset the flags bits in other in the value.

pub fn toggle(&mut self, other: Self)

Toggle the flags in other in the value.

impl ExampleFlags

pub const fn iter(&self) -> Iter<Self>

Yield a set of contained flags values.

Each yielded flags value will correspond to a defined named flag. Any unknown bits will be yielded together as a final flags value.

pub const fn iter_names(&self) -> IterNames<Self>

Yield a set of contained named flags values.

This method is like iter, except only yields bits in contained named flags. Any unknown bits, or bits not corresponding to a contained flag will not be yielded.